requirements.txt
asgiref==3.5.2
asttokens==2.0.8
autopep8==1.7.0
colorama==0.4.5
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
executing==1.1.0
icecream==2.1.3
Pillow==9.2.0
pycodestyle==2.9.1
Pygments==2.13.0
PyJWT==2.5.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
